@@ -1,70 +1,97 @@
 import { InjectionToken, NgZone } from '@angular/core';
 import { isPlatformServer } from '@angular/common';
-import { Observable, Subscription, queueScheduler as queue } from 'rxjs';
+import { Observable, Subscription, SchedulerLike, SchedulerAction, queueScheduler, Operator, Subscriber, TeardownLogic, asyncScheduler } from 'rxjs';
+import { subscribeOn, observeOn, tap, share } from 'rxjs/operators';
 
 // SEMVER put in database.ts when we drop database-depreciated
 export const ɵDATABASE_URL = new InjectionToken<string>('angularfire2.realtimeDatabaseURL');
 
-export class ɵFirebaseZoneScheduler {
-  constructor(public zone: NgZone, private platformId: Object) {}
-  schedule(...args: any[]): Subscription {
-    return <Subscription>this.zone.runGuarded(function() { return queue.schedule.apply(queue, args)});
+function noop() { }
+
+/**
+ * Schedules tasks so that they are invoked inside the Zone that is passed in the constructor.
+ */
+export class ɵZoneScheduler implements SchedulerLike {
+  constructor(private zone: any, private delegate: any = queueScheduler) { }
+
+  now() {
+    return this.delegate.now();
   }
-  // TODO this is a hack, clean it up
-  keepUnstableUntilFirst<T>(obs$: Observable<T>) {
-    if (isPlatformServer(this.platformId)) {
-      return new Observable<T>(subscriber => {
-        const noop = () => {};
-        // @ts-ignore
-        const task = Zone.current.scheduleMacroTask('firebaseZoneBlock', noop, {}, noop, noop);
-        obs$.subscribe(
-          next => {
-            if (task.state === 'scheduled') { task.invoke() };
-            subscriber.next(next);
-          },
-          error => {
-            if (task.state === 'scheduled') { task.invoke() }
-            subscriber.error(error);
-          },
-          () => {
-            if (task.state === 'scheduled') { task.invoke() }
-            subscriber.complete();
-          }
-        );
+
+  schedule(work: (this: SchedulerAction<any>, state?: any) => void, delay?: number, state?: any): Subscription {
+    const targetZone = this.zone;
+    // Wrap the specified work function to make sure that if nested scheduling takes place the
+    // work is executed in the correct zone
+    const workInZone = function (this: SchedulerAction<any>, state: any) {
+      targetZone.runGuarded(() => {
+        work.apply(this, [state]);
       });
-    } else {
-      return obs$;
+    }
+
+    // Scheduling itself needs to be run in zone to ensure setInterval calls for async scheduling are done
+    // inside the correct zone. This scheduler needs to schedule asynchronously always to ensure that
+    // firebase emissions are never synchronous. Specifying a delay causes issues with the queueScheduler delegate.
+    return this.delegate.schedule(workInZone, delay, state)
+  }
+}
+
+export class ɵBlockUntilFirstOperator<T> implements Operator<T, T> {
+  private task: MacroTask | null = null;
+
+  constructor(private zone: any) { }
+
+  call(subscriber: Subscriber<T>, source: Observable<T>): TeardownLogic {
+    const unscheduleTask = this.unscheduleTask.bind(this);
+    this.task = this.zone.run(() => Zone.current.scheduleMacroTask('firebaseZoneBlock', noop, {}, noop, noop));
+
+    return source.pipe(
+      tap(unscheduleTask, unscheduleTask, unscheduleTask)
+    ).subscribe(subscriber).add(unscheduleTask);
+  }
+
+  private unscheduleTask() {
+    if (this.task != null && this.task.state === 'scheduled') {
+      this.task.invoke();
+      this.task = null;
     }
   }
-  runOutsideAngular<T>(obs$: Observable<T>): Observable<T> {
-    return new Observable<T>(subscriber => {
-      return this.zone.runOutsideAngular(() => {
-        return obs$.subscribe(
-          value => this.zone.run(() => subscriber.next(value)),
-          error => this.zone.run(() => subscriber.error(error)),
-          ()    => this.zone.run(() => subscriber.complete()),
-        );
-      });
-    });
+}
+
+export class ɵAngularFireSchedulers {
+  public readonly outsideAngular: ɵZoneScheduler;
+  public readonly insideAngular: ɵZoneScheduler;
+
+  constructor(public ngZone: NgZone) {
+    this.outsideAngular = ngZone.runOutsideAngular(() => new ɵZoneScheduler(Zone.current));
+    this.insideAngular = ngZone.run(() => new ɵZoneScheduler(Zone.current, asyncScheduler));
   }
 }
 
-export const ɵrunOutsideAngular = (zone: NgZone) => <T>(obs$: Observable<T>): Observable<T> => {
-  return new Observable<T>(subscriber => {
-    return zone.runOutsideAngular(() => {
-      ɵrunInZone(zone)(obs$).subscribe(subscriber);
-    });
-  });
-}
+/**
+ * Operator to block the zone until the first value has been emitted or the observable
+ * has completed/errored. This is used to make sure that universal waits until the first
+ * value from firebase but doesn't block the zone forever since the firebase subscription
+ * is still alive.
+ */
+export function ɵkeepUnstableUntilFirstFactory(
+  schedulers: ɵAngularFireSchedulers,
+  platformId: Object
+) {
+  return function keepUnstableUntilFirst<T>(obs$: Observable<T>): Observable<T> {
+    if (isPlatformServer(platformId)) {
+      obs$ = obs$.lift(
+        new ɵBlockUntilFirstOperator(schedulers.ngZone)
+      );
+    }
 
-export const ɵrunInZone = (zone: NgZone) => <T>(obs$: Observable<T>): Observable<T> => {
-  return new Observable<T>(subscriber => {
-    return obs$.subscribe(
-      value => zone.run(() => subscriber.next(value)),
-      error => zone.run(() => subscriber.error(error)),
-      ()    => zone.run(() => subscriber.complete()),
+    return obs$.pipe(
+      // Run the subscribe body outside of Angular (e.g. calling Firebase SDK to add a listener to a change event)
+      subscribeOn(schedulers.outsideAngular),
+      // Run operators inside the angular zone (e.g. side effects via tap())
+      observeOn(schedulers.insideAngular),
+      share()
     );
-  });
+  }
 }
 
 type FunctionPropertyNames<T> = { [K in keyof T]: T[K] extends Function ? K : never }[keyof T];
