@@ -57,11 +57,11 @@ export const projectPrompt = (projects: Project[]) => {
 };
 
 export function getFirebaseProjectName(
-  projectRoot: string,
+  workspaceRoot: string,
   target: string
 ): string | undefined {
   const { targets }: FirebaseRc = JSON.parse(
-    readFileSync(join(projectRoot, ".firebaserc"), "UTF-8")
+    readFileSync(join(workspaceRoot, ".firebaserc"), "UTF-8")
   );
   const projects = Object.keys(targets!);
   return projects.find(
